fix: Add Jina
* fix: Add Jina

I am still running the model

* Add results for Jina

* Add results for Jina

* format fixes

* Update mkdocs to legacy python

---------

+++ src/seb/registered_models/normalize_to_ndarray.py
@@ -10,5 +10,5 @@ def normalize_to_ndarray(embeddings: Union[torch.Tensor, np.ndarray, list[np.nda
             return torch.cat(embeddings).detach().cpu().numpy()  # type: ignore
         return np.concatenate(embeddings)
     if isinstance(embeddings, torch.Tensor):
-        return embeddings.detach().cpu().numpy()
+        return embeddings.detach().cpu().float().numpy()
     return embeddings
